Add fix for sending dates to NPN
We were defaulting to sending the dates with
zeroed out time. This actually affects the date
that is created since the date is sent with no
time zone. We needed to add something in that
parsed the actual hour that's sent over too. In
NPN's case, it's the seventh hour